feat(bin-migration): Task 3 — audit.py + discover.py を scripts/lib/ に移設
- scripts/lib/audit.py: 全 _plugin_root hardcode → PLUGIN_ROOT に置換
- scripts/lib/discover.py: 同上。HISTORY_DIR も PLUGIN_ROOT 基準に修正
- skills/audit/scripts/audit.py: importlib shim (循環インポート回避)
- skills/discover/scripts/discover.py: 同上

モジュール名は変わらないため from audit import run_audit 等は変更不要。
