+++ skills/discover/scripts/discover.py
@@ -1,1094 +1,15 @@
 #!/usr/bin/env python3
-"""パターン発見スクリプト。
-
-usage.jsonl、errors.jsonl、sessions.jsonl、history.jsonl から
-繰り返しパターンを検出し、スキル/ルール候補を生成する。
-"""
-import argparse
-import json
+"""CLI エントリポイント shim — 実装は scripts/lib/discover.py に移設。
+bin/rl-discover が整備されたら SKILL.md から直接呼ばれなくなる。"""
+import importlib.util
 import sys
-from collections import Counter, defaultdict
 from pathlib import Path
-from typing import Any, Dict, List, Optional
 
-DATA_DIR = Path.home() / ".claude" / "rl-anything"
-
-HISTORY_DIR = (
-    Path(__file__).parent.parent
-    / "skills"
-    / "genetic-prompt-optimizer"
-    / "scripts"
-    / "generations"
-)
-
-# 閾値
-BEHAVIOR_THRESHOLD = 5   # 行動パターン検出閾値
-ERROR_THRESHOLD = 3       # エラーパターン検出閾値
-REJECTION_THRESHOLD = 3   # 却下理由検出閾値
-MISSED_SKILL_THRESHOLD = 2  # missed skill 検出閾値（セッション数）
-
-# 構造的制約は共通モジュールから取得
-_plugin_root = Path(__file__).resolve().parent.parent.parent.parent
-sys.path.insert(0, str(_plugin_root / "scripts" / "lib"))
-from agent_classifier import classify_agent_type
-from line_limit import MAX_RULE_LINES, MAX_SKILL_LINES
-from similarity import jaccard_coefficient, tokenize
-from skill_triggers import extract_skill_triggers, normalize_skill_name
-
-# Jaccard 照合閾値
-JACCARD_THRESHOLD = 0.15
-
-# Discover 振動防止用抑制リスト
-SUPPRESSION_FILE = DATA_DIR / "discover-suppression.jsonl"
-
-
-def load_jsonl(filepath: Path) -> List[Dict[str, Any]]:
-    """JSONL ファイルを読み込む。"""
-    if not filepath.exists():
-        return []
-    records = []
-    for line in filepath.read_text(encoding="utf-8").splitlines():
-        try:
-            records.append(json.loads(line))
-        except json.JSONDecodeError:
-            continue
-    return records
-
-
-def load_suppression_list() -> set:
-    """抑制リスト（2回 reject されたパターン）を読み込む。
-
-    type: "merge" エントリは除外し、type 未指定エントリのみを返す。
-    """
-    records = load_jsonl(SUPPRESSION_FILE)
-    return set(r.get("pattern", "") for r in records if r.get("type") != "merge")
-
-
-def load_merge_suppression() -> set:
-    """merge suppression リスト（type: "merge" エントリ）を読み込み、ペアキーの set を返す。"""
-    records = load_jsonl(SUPPRESSION_FILE)
-    return set(r.get("pattern", "") for r in records if r.get("type") == "merge")
-
-
-def add_merge_suppression(skill_a: str, skill_b: str) -> None:
-    """merge suppression エントリを追加する。スキル名をソートし :: 結合で正規化。
-
-    書き込み失敗時は stderr にエラー出力し、例外を送出しない。
-    """
-    key = "::".join(sorted([skill_a, skill_b]))
-    try:
-        DATA_DIR.mkdir(parents=True, exist_ok=True)
-        with open(SUPPRESSION_FILE, "a", encoding="utf-8") as f:
-            f.write(json.dumps({"pattern": key, "type": "merge"}, ensure_ascii=False) + "\n")
-    except OSError as e:
-        print(f"[rl-anything] merge suppression write failed: {e}", file=sys.stderr)
-
-
-def add_to_suppression_list(pattern: str) -> None:
-    """抑制リストにパターンを追加する。"""
-    DATA_DIR.mkdir(parents=True, exist_ok=True)
-    with open(SUPPRESSION_FILE, "a", encoding="utf-8") as f:
-        f.write(json.dumps({"pattern": pattern}, ensure_ascii=False) + "\n")
-
-
-def _load_classify_usage_skill():
-    """audit.py の _is_plugin_skill と classify_usage_skill を遅延インポートで取得する。
-
-    Returns:
-        _is_plugin_skill 関数（classify_usage_skill + _is_gstack_skill + _is_openspec_skill の併用）
-    """
-    import sys as _sys
-    _plugin_root = Path(__file__).resolve().parent.parent.parent.parent
-    _audit_scripts = _plugin_root / "skills" / "audit" / "scripts"
-    if str(_audit_scripts) not in _sys.path:
-        _sys.path.insert(0, str(_audit_scripts))
-    from audit import _is_plugin_skill, classify_usage_skill
-    return _is_plugin_skill, classify_usage_skill
-
-
-def detect_behavior_patterns(
-    threshold: int = BEHAVIOR_THRESHOLD,
-    *,
-    project_root: Optional[Path] = None,
-    include_unknown: bool = False,
-) -> List[Dict[str, Any]]:
-    """繰り返し行動パターンの検出（usage + sessions、5+閾値）。
-
-    parent_skill の有無で contextualized / ad-hoc / unknown を分類し、
-    ad-hoc パターンのみをスキル候補として提案する。
-    処理順序:
-    1. プラグインスキルはメインランキングから除外し、plugin_summary に集約
-    2. 組み込み Agent は agent_usage_summary に分離
-    3. カスタム Agent はメインランキングに残留
-    """
-    from telemetry_query import query_usage
-
-    project_name = project_root.name if project_root else None
-    usage = query_usage(
-        project=project_name,
-        include_unknown=include_unknown,
-        usage_file=DATA_DIR / "usage.jsonl",
-    )
-    _is_plugin, _classify = _load_classify_usage_skill()
-
-    # ad-hoc レコードのみカウント（contextualized/unknown を除外）
-    ad_hoc_counter: Counter = Counter()
-    all_counter: Counter = Counter()
-    plugin_counter: Counter = Counter()  # プラグイン別集計
-    for rec in usage:
-        skill = rec.get("skill_name", "")
-        if not skill:
-            continue
-        all_counter[skill] += 1
-
-        # (1) プラグインスキルは別集計
-        if _is_plugin(skill):
-            plugin_name = _classify(skill) or "gstack"
-            plugin_counter[plugin_name] += 1
-            continue
-
-        parent_skill = rec.get("parent_skill")
-        source = rec.get("source", "")
-
-        # backfill データ（parent_skill なし + source=backfill）は unknown として除外
-        if parent_skill is None and source == "backfill":
-            continue
-        # contextualized（parent_skill あり）は除外
-        if parent_skill is not None:
-            continue
-        # ad-hoc（parent_skill なし、backfill でない）のみカウント
-        ad_hoc_counter[skill] += 1
-
-    suppressed = load_suppression_list()
-    patterns = []
-    builtin_agent_counter: Counter = Counter()
-    builtin_agent_prompts: Dict[str, List[str]] = defaultdict(list)
-
-    for skill, ad_hoc_count in ad_hoc_counter.most_common():
-        if ad_hoc_count < threshold or skill in suppressed:
-            continue
-
-        # (2) Agent:XX パターンの分類
-        if skill.startswith("Agent:"):
-            agent_name = skill[len("Agent:"):]
-            agent_type = classify_agent_type(agent_name, project_root=project_root)
-
-            prompts = [
-                r.get("prompt", "") for r in usage
-                if r.get("skill_name") == skill
-                and r.get("prompt")
-                and r.get("parent_skill") is None
-                and r.get("source", "") != "backfill"
-            ]
-
-            if agent_type == "builtin":
-                # 組み込み Agent → builtin_agent_counter に分離
-                builtin_agent_counter[skill] = ad_hoc_count
-                builtin_agent_prompts[skill] = prompts
-                continue
-
-            # (3) カスタム Agent → メインランキングに残留
-            pattern: Dict[str, Any] = {
-                "type": "behavior",
-                "pattern": skill,
-                "count": ad_hoc_count,
-                "total_count": all_counter.get(skill, 0),
-                "suggestion": "skill_candidate",
-                "agent_type": agent_type,
-            }
-            subcategories = _classify_agent_prompts(prompts)
-            if subcategories:
-                pattern["subcategories"] = subcategories
-            patterns.append(pattern)
-            continue
-
-        # 非 Agent パターン
-        pattern = {
-            "type": "behavior",
-            "pattern": skill,
-            "count": ad_hoc_count,
-            "total_count": all_counter.get(skill, 0),
-            "suggestion": "skill_candidate",
-        }
-        patterns.append(pattern)
-
-    # プラグイン利用サマリを末尾に付加（保護状態を表示）
-    if plugin_counter:
-        patterns.append({
-            "type": "plugin_summary",
-            "pattern": "plugin_usage",
-            "count": sum(plugin_counter.values()),
-            "suggestion": "info_only",
-            "plugin_breakdown": dict(plugin_counter.most_common()),
-            "protected": True,
-        })
-
-    # 組み込み Agent 利用サマリを末尾に付加
-    if builtin_agent_counter:
-        agent_breakdown: Dict[str, Any] = {}
-        for agent_skill, count in builtin_agent_counter.most_common():
-            entry: Dict[str, Any] = {"count": count}
-            subcategories = _classify_agent_prompts(builtin_agent_prompts.get(agent_skill, []))
-            if subcategories:
-                entry["subcategories"] = subcategories
-            agent_breakdown[agent_skill] = entry
-
-        patterns.append({
-            "type": "agent_usage_summary",
-            "pattern": "builtin_agent_usage",
-            "count": sum(builtin_agent_counter.values()),
-            "suggestion": "info_only",
-            "agent_breakdown": agent_breakdown,
-        })
-
-    return patterns
-
-
-def _classify_agent_prompts(prompts: List[str]) -> List[Dict[str, Any]]:
-    """Agent の prompt リストをキーワードベースで簡易分類する。
-
-    common.PROMPT_CATEGORIES / common.classify_prompt() を利用。
-    """
-    # hooks/common.py をインポート
-    import sys as _sys
-    _plugin_root = Path(__file__).resolve().parent.parent.parent.parent
-    if str(_plugin_root / "hooks") not in _sys.path:
-        _sys.path.insert(0, str(_plugin_root / "hooks"))
-    import common as _common
-
-    category_counts: Counter = Counter()
-    category_examples: Dict[str, str] = {}
-
-    for prompt in prompts:
-        cat = _common.classify_prompt(prompt)
-        category_counts[cat] += 1
-        if cat != "other" and cat not in category_examples:
-            category_examples[cat] = prompt[:120]
-
-    results = []
-    for cat, count in category_counts.most_common():
-        entry: Dict[str, Any] = {
-            "category": cat,
-            "count": count,
-        }
-        if cat in category_examples:
-            entry["example"] = category_examples[cat]
-        results.append(entry)
-    return results
-
-
-def detect_missed_skills(
-    threshold: int = MISSED_SKILL_THRESHOLD,
-    *,
-    project_root: Optional[Path] = None,
-    include_unknown: bool = False,
-) -> Dict[str, Any]:
-    """スキルのトリガーワードにマッチしたがスキルが使われなかったセッションを検出する。
-
-    Returns:
-        {"missed": [...], "message": str or None}
-        missed: [{"skill": str, "triggers_matched": [str], "session_count": int}, ...]
-    """
-    from telemetry_query import query_sessions, query_usage
-
-    # CLAUDE.md からスキルトリガーを取得
-    skill_triggers = extract_skill_triggers(project_root=project_root)
-    if not skill_triggers:
-        return {"missed": [], "message": "No CLAUDE.md found, skipping missed skill detection"}
-
-    project_name = project_root.name if project_root else None
-
-    # sessions.jsonl からセッションデータを取得
-    sessions_file = DATA_DIR / "sessions.jsonl"
-    sessions = query_sessions(
-        project=project_name,
-        include_unknown=include_unknown,
-        sessions_file=sessions_file,
-    )
-    if not sessions:
-        if not sessions_file.exists():
-            return {"missed": [], "message": "No sessions.jsonl found (run backfill first), skipping missed skill detection"}
-        return {"missed": [], "message": None}
-
-    # usage.jsonl からスキル使用実績を取得
-    usage = query_usage(
-        project=project_name,
-        include_unknown=include_unknown,
-        usage_file=DATA_DIR / "usage.jsonl",
-    )
-
-    # session_id ごとに使用されたスキル名を集約
-    used_skills_by_session: Dict[str, set] = defaultdict(set)
-    for rec in usage:
-        sid = rec.get("session_id", "")
-        skill = rec.get("skill_name", "")
-        if sid and skill:
-            used_skills_by_session[sid].add(normalize_skill_name(skill))
-
-    # セッションごとにトリガーマッチ → スキル使用チェック
-    # skill -> {triggers_matched: set, sessions: set}
-    missed_map: Dict[str, Dict[str, Any]] = defaultdict(lambda: {"triggers_matched": set(), "sessions": set()})
-
-    for session in sessions:
-        sid = session.get("session_id", "")
-        user_prompts = session.get("user_prompts", [])
-        if not sid or not user_prompts:
-            continue
-
-        prompts_text = " ".join(user_prompts).lower()
-        used_in_session = used_skills_by_session.get(sid, set())
-
-        for skill_entry in skill_triggers:
-            skill_name = skill_entry["skill"]
-            if skill_name in used_in_session:
-                continue
-
-            for trigger in skill_entry["triggers"]:
-                if trigger.lower() in prompts_text:
-                    missed_map[skill_name]["triggers_matched"].add(trigger)
-                    missed_map[skill_name]["sessions"].add(sid)
-
-    # 閾値フィルタリング
-    missed = []
-    for skill, data in sorted(missed_map.items(), key=lambda x: len(x[1]["sessions"]), reverse=True):
-        session_count = len(data["sessions"])
-        if session_count >= threshold:
-            entry = {
-                "skill": skill,
-                "triggers_matched": sorted(data["triggers_matched"]),
-                "session_count": session_count,
-            }
-            # eval set ステータスを付与
-            eval_set_path = Path.home() / ".claude" / "rl-anything" / "eval-sets" / f"{skill}.json"
-            if eval_set_path.exists():
-                entry["eval_set_path"] = str(eval_set_path)
-                entry["eval_set_status"] = "available"
-            else:
-                entry["eval_set_path"] = None
-                entry["eval_set_status"] = "not_generated"
-            missed.append(entry)
-
-    return {"missed": missed, "message": None}
-
-
-def detect_error_patterns(
-    threshold: int = ERROR_THRESHOLD,
-    *,
-    project_root: Optional[Path] = None,
-    include_unknown: bool = False,
-) -> List[Dict[str, Any]]:
-    """繰り返しエラーパターンの検出（errors、3+閾値）。"""
-    from telemetry_query import query_errors
-
-    project_name = project_root.name if project_root else None
-    errors = query_errors(
-        project=project_name,
-        include_unknown=include_unknown,
-        errors_file=DATA_DIR / "errors.jsonl",
-    )
-    counter: Counter = Counter()
-    for rec in errors:
-        error = rec.get("error", "")[:200]
-        if error:
-            counter[error] += 1
-
-    suppressed = load_suppression_list()
-    patterns = []
-    for error, count in counter.most_common():
-        if count >= threshold and error not in suppressed:
-            patterns.append({
-                "type": "error",
-                "pattern": error,
-                "count": count,
-                "suggestion": "rule_candidate",
-            })
-    return patterns
-
-
-def detect_rejection_patterns(threshold: int = REJECTION_THRESHOLD) -> List[Dict[str, Any]]:
-    """繰り返し却下理由の検出（rejection_reason、3+閾値）。"""
-    history_file = HISTORY_DIR / "history.jsonl"
-    records = load_jsonl(history_file)
-
-    counter: Counter = Counter()
-    for rec in records:
-        reason = rec.get("rejection_reason")
-        if reason:
-            counter[reason] += 1
-
-    suppressed = load_suppression_list()
-    patterns = []
-    for reason, count in counter.most_common():
-        if count >= threshold and reason not in suppressed:
-            patterns.append({
-                "type": "rejection",
-                "pattern": reason,
-                "count": count,
-                "suggestion": "rule_candidate",
-            })
-    return patterns
-
-
-def determine_scope(pattern: Dict[str, Any]) -> str:
-    """スコープ配置の判断（global / project / plugin）。
-
-    カスタム Agent は agent_type フィールドから判定する。
-    """
-    # カスタム Agent: agent_type フィールドで判定
-    agent_type = pattern.get("agent_type")
-    if agent_type == "custom_global":
-        return "global"
-    if agent_type == "custom_project":
-        return "project"
-
-    p = pattern.get("pattern", "").lower()
-
-    # global 配置の兆候
-    global_keywords = ["git", "commit", "pr", "pull request", "test", "lint", "claude", "format"]
-    if any(kw in p for kw in global_keywords):
-        return "global"
-
-    # project 配置の兆候
-    project_keywords = ["react", "next", "vue", "django", "rails", "prisma", "docker"]
-    if any(kw in p for kw in project_keywords):
-        return "project"
-
-    return "project"  # デフォルト
-
-
-def validate_skill_content(content: str) -> bool:
-    """スキル候補の構造バリデーション（MUST 500行以下）。"""
-    lines = content.count("\n") + 1
-    return lines <= MAX_SKILL_LINES
-
-
-def validate_rule_content(content: str) -> bool:
-    """ルール候補の構造バリデーション（MUST 3行以内）。"""
-    lines = content.count("\n") + 1
-    return lines <= MAX_RULE_LINES
-
-
-def load_claude_reflect_data() -> List[Dict[str, Any]]:
-    """corrections.jsonl から pending の修正データのみ取り込む。未生成時はスキップ。
-
-    reflect が処理するのは pending のみであるため、
-    evolve の reflect_data_count と reflect の認識を一致させる。
-    """
-    corrections_file = DATA_DIR / "corrections.jsonl"
-
-    if not corrections_file.exists():
-        return []
-
-    records = load_jsonl(corrections_file)
-    return [r for r in records if r.get("reflect_status", "pending") == "pending"]
-
-
-# ---------- enrich (Jaccard 照合、旧 enrich.py から統合) ----------
-
-
-def _load_skill_tokens(skill_path: Path) -> Dict[str, Any]:
-    """SKILL.md の先頭 50 行 + スキル名からトークン集合を生成する。"""
-    from typing import Set as _Set
-
-    tokens: _Set[str] = set()
-    skill_name = skill_path.parent.name
-    tokens |= tokenize(skill_name)
-
-    try:
-        lines = skill_path.read_text(encoding="utf-8").splitlines()[:50]
-        for line in lines:
-            tokens |= tokenize(line)
-    except OSError:
-        pass
-
-    return {"path": skill_path, "name": skill_name, "tokens": tokens}
-
-
-def _enrich_patterns(
-    patterns: List[Dict[str, Any]],
-    project_dir: Optional[Path] = None,
-    max_matches: int = 3,
-) -> Dict[str, Any]:
-    """パターンを既存スキルに Jaccard 係数でマッチングする。
-
-    enrich.py から統合。プラグイン由来のスキルは除外。
-    Jaccard >= JACCARD_THRESHOLD のマッチのみ保持し、上位 max_matches 件を返す。
-
-    Returns:
-        {"matched_skills": [...], "unmatched_patterns": [...]}
-    """
-    proj = project_dir or Path.cwd()
-
-    # audit.py から find_artifacts / classify_artifact_origin を遅延インポート
-    import sys as _sys
-    _audit_scripts = _plugin_root / "skills" / "audit" / "scripts"
-    if str(_audit_scripts) not in _sys.path:
-        _sys.path.insert(0, str(_audit_scripts))
-    from audit import classify_artifact_origin, find_artifacts
-
-    artifacts = find_artifacts(proj)
-
-    # 非プラグインスキルのトークンを事前計算
-    skill_info = []
-    for skill_path in artifacts.get("skills", []):
-        origin = classify_artifact_origin(skill_path)
-        if origin == "plugin":
-            continue
-        skill_info.append(_load_skill_tokens(skill_path))
-
-    matched_skills: List[Dict[str, Any]] = []
-    matched_pattern_texts: set = set()
-
-    for pattern in patterns:
-        pattern_text = pattern.get("pattern", "")
-        pattern_type = pattern.get("type", "unknown")
-        pattern_tokens = tokenize(pattern_text)
-
-        if not pattern_tokens:
-            continue
-
-        scored = []
-        for info in skill_info:
-            score = jaccard_coefficient(pattern_tokens, info["tokens"])
-            if score >= JACCARD_THRESHOLD:
-                scored.append({
-                    "pattern_type": pattern_type,
-                    "pattern": pattern_text,
-                    "matched_skill": info["name"],
-                    "skill_path": str(info["path"]),
-                    "jaccard_score": round(score, 4),
-                })
-
-        scored.sort(key=lambda x: x["jaccard_score"], reverse=True)
-        if scored:
-            matched_skills.extend(scored[:max_matches])
-            matched_pattern_texts.add(pattern_text)
-
-    # 未マッチパターン
-    unmatched_patterns = []
-    for pattern in patterns:
-        pattern_text = pattern.get("pattern", "")
-        if pattern_text not in matched_pattern_texts:
-            unmatched_patterns.append({
-                "pattern_type": pattern.get("type", "unknown"),
-                "pattern": pattern_text,
-                "suggestion": pattern.get("suggestion", "skill_candidate"),
-            })
-
-    return {
-        "matched_skills": matched_skills,
-        "unmatched_patterns": unmatched_patterns,
-    }
-
-
-# ---------- recommended artifacts ----------
-
-RECOMMENDED_ARTIFACTS = [
-    {
-        "id": "no-defer-use-subagent",
-        "type": "rule",
-        "path": Path.home() / ".claude" / "rules" / "no-defer-use-subagent.md",
-        "description": "先送り禁止 — background subagent 即時委譲ルール",
-        "hook_path": Path.home() / ".claude" / "hooks" / "detect-deferred-task.py",
-        "hook_description": "Stop hook: 先送り表現検出 → 会話続行強制",
-    },
-    {
-        "id": "avoid-bash-builtin",
-        "type": "rule+hook",
-        "path": Path.home() / ".claude" / "rules" / "avoid-bash-builtin.md",
-        "description": "Bash Built-in 代替コマンド禁止 — grep/cat/find 等を Built-in ツールに誘導",
-        "hook_path": Path.home() / ".claude" / "hooks" / "check-bash-builtin.py",
-        "hook_description": "PreToolUse hook: Bash で Built-in 代替可能コマンドを block",
-        "data_driven": True,  # tool_usage_patterns のデータで提案根拠を補強
-        "recommendation_id": "builtin_replaceable",
-        "content_patterns": ["REPLACEABLE"],
-    },
-    {
-        "id": "sleep-polling-guard",
-        "type": "hook",
-        "path": None,
-        "description": "sleep ポーリング検出 — run_in_background + 完了通知待ちを推奨",
-        "hook_path": Path.home() / ".claude" / "hooks" / "check-bash-builtin.py",
-        "hook_description": "PreToolUse hook: sleep コマンドを検出・警告",
-        "recommendation_id": "sleep_polling",
-        "content_patterns": [r"\bsleep\b"],
-    },
-    {
-        "id": "test-happy-path-first",
-        "type": "rule",
-        "path": Path.home() / ".claude" / "rules" / "test-happy-path-first.md",
-        "description": "テストはハッピーパスから書く — パイプラインの正常系E2Eテストを最初に書くルール",
-        "hook_path": None,
-    },
-    {
-        "id": "commit-version",
-        "type": "rule",
-        "path": Path.home() / ".claude" / "rules" / "commit-version.md",
-        "description": "コミット時バージョン管理 — feat!=major, feat=minor, fix=patch の自動判定提案",
-        "hook_path": None,
-    },
-    {
-        "id": "commit-skill",
-        "type": "skill",
-        "path": Path.home() / ".claude" / "skills" / "commit" / "SKILL.md",
-        "description": "Conventional Commits + CHANGELOG 自動追記スキル — feat→Added, fix→Fixed, BREAKING CHANGE 対応",
-        "hook_path": None,
-    },
-    {
-        "id": "claude-md-style",
-        "type": "rule",
-        "path": Path.home() / ".claude" / "rules" / "claude-md-style.md",
-        "description": "CLAUDE.md vs MEMORY.md の使い分け — 動作情報/経緯/コードから分かる情報の記載先判断",
-        "hook_path": None,
-    },
-    {
-        "id": "process-stall-guard",
-        "type": "rule",
-        "path": Path.home() / ".claude" / "rules" / "process-stall-guard.md",
-        "description": "長時間プロセス実行前に既存プロセスを確認 — CDK/Docker/npm 等の停滞→kill→リトライ防止",
-        "hook_path": None,
-        "recommendation_id": "process_stall_guard",
-        "content_patterns": [r"\bpgrep\b", r"\bkill\b", r"stall|停滞"],
-    },
-    {
-        "id": "evidence-before-claims",
-        "type": "rule",
-        "path": Path.home() / ".claude" / "rules" / "verify-before-claim.md",
-        "description": "証拠提示義務 — 完了主張の前に検証コマンドの実行結果を提示する",
-        "hook_path": None,
-        "recommendation_id": "evidence_before_claims",
-        "content_patterns": ["verify-before-claim", "evidence", "証拠", r"完了.*確認"],
-    },
-    # --- 構造化実装 ---
-    {
-        "id": "implement-skill",
-        "type": "skill",
-        "path": Path(__file__).resolve().parents[2] / "implement" / "SKILL.md",
-        "description": "implement — plan artifact → タスク分解 → 実装（Standard/Parallel）→ 計画準拠チェック → テレメトリ記録",
-        "hook_path": None,
-    },
-    {
-        "id": "suggest-implement-skill",
-        "type": "rule",
-        "path": Path.home() / ".claude" / "rules" / "suggest-implement-skill.md",
-        "description": "実装タスク時の implement スキル提案 — 「実装して」等で /rl-anything:implement を提案",
-        "hook_path": None,
-    },
-    {
-        "id": "implement-flow-chain",
-        "type": "config",
-        "path": Path.home() / ".gstack" / "flow-chain.json",
-        "description": "gstack フローチェーンに implement を追加 — plan-eng-review → /rl-anything:implement → /review",
-        "hook_path": None,
-        "content_patterns": ["implement"],
-    },
-    # --- gstack ワークフローツール ---
-    {
-        "id": "gstack-flow-chain",
-        "type": "rule",
-        "path": Path.home() / ".claude" / "rules" / "gstack-flow-chain.md",
-        "description": "gstack フローチェーン — /ship→/document-release→/spec-keeper update→/retro の実装後ワークフロー",
-        "hook_path": None,
-    },
-    {
-        "id": "living-spec-awareness",
-        "type": "rule",
-        "path": Path.home() / ".claude" / "rules" / "living-spec-awareness.md",
-        "description": "Living Spec 意識 — SPEC.md 未存在 PJ で /spec-keeper init を提案、存在 PJ では最初に読む",
-        "hook_path": None,
-    },
-    {
-        "id": "spec-keeper",
-        "type": "skill",
-        "path": Path(__file__).resolve().parents[2] / "spec-keeper" / "SKILL.md",
-        "description": "spec-keeper — SPEC.md + ADR 管理スキル（init/update/adr/status）",
-        "hook_path": None,
-    },
-    {
-        "id": "ship",
-        "type": "skill",
-        "path": Path.home() / ".claude" / "skills" / "ship" / "SKILL.md",
-        "description": "ship — 実装→テスト→bump→CHANGELOG→PR の出荷ワークフロー",
-        "hook_path": None,
-    },
-    {
-        "id": "continuation-check",
-        "type": "rule",
-        "path": Path.home() / ".claude" / "rules" / "continuation-check.md",
-        "description": "前回の続き判定 — handover/ロードマップ言及時に引き継ぎファイルとSPEC.mdを自動確認",
-        "hook_path": None,
-    },
-    # --- 並行開発パターン ---
-    {
-        "id": "worktree-parallel-work",
-        "type": "rule+hook",
-        "path": Path.home() / ".claude" / "rules" / "worktree-parallel-work.md",
-        "description": "worktree 並行開発 — ブランチ作成時に worktree を使い、stash+checkout 事故と同一ディレクトリ並行作業を防止。feature-branch rule の PJ 上書きが必要",
-        "hook_path": Path.home() / ".claude" / "hooks" / "check-worktree.py",
-        "hook_description": "PreToolUse hook: git stash+checkout および git checkout -b を検出し worktree を提案",
-    },
-    {
-        "id": "deploy-lock",
-        "type": "hook",
-        "path": None,
-        "description": "デプロイ排他制御 — 同一環境への並行デプロイを lock ファイルで防止（PreToolUse で取得 + PostToolUse で解放）",
-        "hook_path": Path.home() / ".claude" / "hooks" / "deploy-lock.py",
-        "hook_description": "PreToolUse hook: deploy 時に環境別 lock 取得。PostToolUse hook: deploy 完了時に lock 解放",
-    },
-    {
-        "id": "kill-guard",
-        "type": "hook",
-        "path": None,
-        "description": "プロセス kill ガード — deploy-lock 保持中のプロセス kill をブロック",
-        "hook_path": Path.home() / ".claude" / "hooks" / "kill-guard.py",
-        "hook_description": "PreToolUse hook: kill/pkill/pgrep+kill が deploy 関連プロセスを対象とする場合、active lock があればブロック",
-    },
-]
-
-
-def detect_recommended_artifacts(
-    tool_usage_patterns: Optional[Dict[str, Any]] = None,
-) -> List[Dict[str, Any]]:
-    """推奨ルール/hook の導入状態をチェックする。
-
-    Returns:
-        未導入 artifact のリスト。各エントリに evidence (検出データ) を含む。
-    """
-    missing = []
-    for artifact in RECOMMENDED_ARTIFACTS:
-        rule_path = artifact.get("path")
-        rule_exists = rule_path.exists() if rule_path else True
-        hook_path = artifact.get("hook_path")
-        hook_exists = hook_path.exists() if hook_path else True
-
-        if rule_exists and hook_exists:
-            continue
-
-        entry: Dict[str, Any] = {
-            "id": artifact["id"],
-            "description": artifact["description"],
-            "missing": [],
-        }
-        if not rule_exists and rule_path:
-            entry["missing"].append({"type": "rule", "path": str(rule_path)})
-        if not hook_exists and hook_path:
-            entry["missing"].append({
-                "type": "hook",
-                "path": str(hook_path),
-                "description": artifact.get("hook_description", ""),
-            })
-
-        # data_driven な artifact には tool_usage データを証拠として付加
-        if artifact.get("data_driven") and tool_usage_patterns:
-            builtin = tool_usage_patterns.get("builtin_replaceable", [])
-            if builtin:
-                entry["evidence"] = {
-                    "builtin_replaceable_count": sum(
-                        item.get("count", 0) for item in builtin
-                    ),
-                    "top_patterns": builtin[:5],
-                }
-                # rule/hook 候補も付加
-                if "rule_candidates" in tool_usage_patterns:
-                    entry["rule_candidates"] = tool_usage_patterns["rule_candidates"]
-                if "hook_candidate" in tool_usage_patterns:
-                    entry["hook_candidate"] = tool_usage_patterns["hook_candidate"]
-
-        missing.append(entry)
-    return missing
-
-
-def detect_installed_artifacts(
-    tool_usage_patterns: Optional[Dict[str, Any]] = None,
-) -> List[Dict[str, Any]]:
-    """導入済み推奨 artifact の状態を返す。
-
-    Returns:
-        導入済み artifact のリスト。hook_status 含む。
-        recommendation_id 付きエントリには mitigation_metrics を付加。
-    """
-    from tool_usage_analyzer import check_artifact_installed
-
-    installed = []
-    for artifact in RECOMMENDED_ARTIFACTS:
-        check_result = check_artifact_installed(artifact)
-        if not check_result["installed"]:
-            continue
-
-        entry: Dict[str, Any] = {
-            "id": artifact["id"],
-            "description": artifact["description"],
-            "status": "active",
-        }
-
-        # data_driven な artifact には hook_status を付加
-        if artifact.get("data_driven") and tool_usage_patterns:
-            hook_status = tool_usage_patterns.get("hook_status")
-            if hook_status:
-                entry["hook_status"] = hook_status
-
-        # recommendation_id 付きエントリには mitigation_metrics を付加
-        rec_id = artifact.get("recommendation_id")
-        if rec_id:
-            entry["recommendation_id"] = rec_id
-            metrics = _compute_mitigation_metrics(
-                rec_id, tool_usage_patterns, check_result,
-            )
-            entry["mitigation_metrics"] = metrics
-
-        installed.append(entry)
-    return installed
-
-
-def _compute_mitigation_metrics(
-    recommendation_id: str,
-    tool_usage_patterns: Optional[Dict[str, Any]],
-    check_result: Dict[str, Any],
-) -> Dict[str, Any]:
-    """recommendation_id に応じた条件別メトリクスを算出する。"""
-    metrics: Dict[str, Any] = {
-        "mitigated": True,
-        "recent_count": 0,
-        "content_matched": check_result.get("content_matched"),
-    }
-
-    if not tool_usage_patterns:
-        return metrics
-
-    if recommendation_id == "builtin_replaceable":
-        builtin = tool_usage_patterns.get("builtin_replaceable", [])
-        metrics["recent_count"] = sum(
-            item.get("count", 0) for item in builtin
-        )
-    elif recommendation_id == "sleep_polling":
-        repeating = tool_usage_patterns.get("repeating_patterns", [])
-        metrics["recent_count"] = sum(
-            item.get("count", 0) for item in repeating
-            if "sleep" in item.get("pattern", "").lower()
-        )
-
-    return metrics
-
-
-def run_discover(
-    *,
-    project_root: Optional[Path] = None,
-    include_unknown: bool = False,
-    tool_usage: bool = False,
-) -> Dict[str, Any]:
-    """Discover を実行して候補を返す。enrich 統合済み。"""
-    behavior = detect_behavior_patterns(
-        project_root=project_root, include_unknown=include_unknown,
-    )
-    errors = detect_error_patterns(
-        project_root=project_root, include_unknown=include_unknown,
-    )
-    rejections = detect_rejection_patterns()
-    reflect_data = load_claude_reflect_data()
-
-    # missed skill 検出
-    missed_result = detect_missed_skills(
-        project_root=project_root,
-        include_unknown=include_unknown,
-    )
-
-    result: Dict[str, Any] = {
-        "behavior_patterns": behavior,
-        "error_patterns": errors,
-        "rejection_patterns": rejections,
-        "reflect_data_count": len(reflect_data),
-    }
-
-    # missed skill opportunities をレポートに含める
-    if missed_result["missed"]:
-        result["missed_skill_opportunities"] = missed_result["missed"]
-    if missed_result["message"]:
-        result["missed_skill_message"] = missed_result["message"]
-
-    # スコープ判断
-    all_patterns = behavior + errors + rejections
-    for p in all_patterns:
-        p["scope"] = determine_scope(p)
-
-    result["total_candidates"] = len(all_patterns)
-
-    # enrich 統合: Jaccard 照合
-    active_patterns = errors + rejections if (errors or rejections) else behavior
-    if active_patterns:
-        enrich_result = _enrich_patterns(active_patterns, project_dir=project_root)
-        result["matched_skills"] = enrich_result["matched_skills"]
-        result["unmatched_patterns"] = enrich_result["unmatched_patterns"]
-
-    # 検証知見カタログの検出
-    try:
-        from verification_catalog import detect_verification_needs
-        proj = project_root or Path.cwd()
-        verification_needs = detect_verification_needs(proj)
-        if verification_needs:
-            result["verification_needs"] = verification_needs
-    except Exception as e:
-        result["verification_needs_error"] = str(e)
-
-    tool_result = None
-    if tool_usage:
-        from tool_usage_analyzer import analyze_tool_usage
-        tool_result = analyze_tool_usage(project_root=project_root)
-        if tool_result["total_tool_calls"] > 0:
-            result["tool_usage_patterns"] = tool_result
-
-    # 推奨アーティファクト未導入チェック（tool_usage データを証拠として付加）
-    recommended_missing = detect_recommended_artifacts(
-        tool_usage_patterns=tool_result,
-    )
-    if recommended_missing:
-        result["recommended_artifacts"] = recommended_missing
-
-    # 導入済みアーティファクトの状態
-    installed = detect_installed_artifacts(
-        tool_usage_patterns=tool_result,
-    )
-    if installed:
-        result["installed_artifacts"] = installed
-
-    # pitfall 自動検出
-    try:
-        _lib_path = Path(__file__).resolve().parent.parent.parent.parent / "scripts" / "lib"
-        if str(_lib_path) not in sys.path:
-            sys.path.insert(0, str(_lib_path))
-        from pitfall_manager import extract_pitfall_candidates
-        from telemetry_query import query_corrections, query_errors
-        proj = project_root or Path.cwd()
-        proj_name = proj.name
-        corrections_data = query_corrections(project=proj_name)
-        errors_data = query_errors(project=proj_name)
-        pitfall_result = extract_pitfall_candidates(corrections_data, errors=errors_data)
-        if pitfall_result["candidates"]:
-            result["pitfall_candidates"] = pitfall_result["candidates"]
-    except Exception as e:
-        result["pitfall_candidates_error"] = str(e)
-
-    # instruction violation 検出 (issue #39)
-    try:
-        from critical_instruction_extractor import (
-            extract_critical_lines,
-            detect_instruction_violation,
-        )
-        from telemetry_query import query_corrections
-        from issue_schema import make_instruction_violation_issue
-
-        proj = project_root or Path.cwd()
-        proj_name = proj.name
-        corrections_data = query_corrections(project=proj_name)
-
-        # last_skill が設定されている corrections のみ対象
-        skill_corrections = [
-            c for c in corrections_data if c.get("last_skill")
-        ]
-
-        violations = []
-        for corr in skill_corrections:
-            skill_name = corr["last_skill"]
-            # スキルの SKILL.md を探す
-            skill_dirs = list(Path.home().glob(f".claude/skills/{skill_name}/SKILL.md"))
-            pj_skill_dirs = list((proj / ".claude" / "skills" / skill_name / "SKILL.md").parent.glob("SKILL.md")) if (proj / ".claude" / "skills" / skill_name).exists() else []
-            all_skill_mds = skill_dirs + [d for d in pj_skill_dirs if d not in skill_dirs]
-
-            for skill_md in all_skill_mds:
-                content = skill_md.read_text(encoding="utf-8")
-                instructions = extract_critical_lines(content)
-                if not instructions:
-                    continue
-                violation = detect_instruction_violation(corr, instructions)
-                if violation:
-                    violations.append(
-                        make_instruction_violation_issue(
-                            skill_name=skill_name,
-                            skill_path=str(skill_md),
-                            instruction_text=violation.instruction.original,
-                            correction_message=violation.correction_message,
-                            match_type=violation.match_type,
-                            confidence=violation.confidence,
-                            reason=violation.reason,
-                            needs_review=violation.needs_review,
-                        )
-                    )
-                break  # 最初にマッチしたスキルのみ
-
-        if violations:
-            result["instruction_violations"] = violations
-    except Exception as e:
-        result["instruction_violations_error"] = str(e)
-
-    # 停滞→リカバリパターン検出
-    try:
-        from tool_usage_analyzer import (
-            extract_tool_calls_by_session,
-            detect_stall_recovery_patterns,
-            STALL_RECOVERY_RECENCY_DAYS,
-        )
-        session_commands = extract_tool_calls_by_session(
-            project_root,
-            max_age_days=STALL_RECOVERY_RECENCY_DAYS,
-        )
-        stall_patterns = detect_stall_recovery_patterns(session_commands)
-        result["stall_recovery_patterns"] = stall_patterns
-    except Exception as e:
-        result["stall_recovery_patterns"] = []
-        result["stall_recovery_error"] = str(e)
-
-    # ワークフローチェックポイントギャップ走査
-    try:
-        from workflow_checkpoint import is_workflow_skill, detect_checkpoint_gaps
-        proj = project_root or Path.cwd()
-        skills_dir = proj / ".claude" / "skills"
-        workflow_gaps = []
-        if skills_dir.is_dir():
-            for skill_dir in sorted(skills_dir.iterdir()):
-                if not skill_dir.is_dir():
-                    continue
-                if not is_workflow_skill(skill_dir):
-                    continue
-                gaps = detect_checkpoint_gaps(skill_dir.name, skill_dir, proj)
-                if gaps:
-                    workflow_gaps.append({
-                        "skill_name": skill_dir.name,
-                        "gaps": gaps,
-                    })
-        if workflow_gaps:
-            result["workflow_checkpoint_gaps"] = workflow_gaps
-    except Exception as e:
-        result["workflow_checkpoint_gaps_error"] = str(e)
-
-    return result
-
-
-def main() -> None:
-    parser = argparse.ArgumentParser(description="パターン発見スクリプト")
-    parser.add_argument(
-        "--project-dir",
-        default=None,
-        help="プロジェクトディレクトリ（指定時はそのプロジェクトのレコードのみ集計）",
-    )
-    parser.add_argument(
-        "--include-unknown",
-        action="store_true",
-        help="project が null のレコードも集計に含める",
-    )
-    parser.add_argument(
-        "--tool-usage",
-        action="store_true",
-        help="セッション JSONL からツール利用パターンを分析する",
-    )
-    args = parser.parse_args()
-
-    project_root = Path(args.project_dir) if args.project_dir else None
-    result = run_discover(
-        project_root=project_root,
-        include_unknown=args.include_unknown,
-        tool_usage=args.tool_usage,
-    )
-    print(json.dumps(result, ensure_ascii=False, indent=2))
-
+_lib = Path(__file__).resolve().parent.parent.parent.parent / "scripts" / "lib"
+_spec = importlib.util.spec_from_file_location("discover", _lib / "discover.py")
+_mod = importlib.util.module_from_spec(_spec)
+sys.modules["discover"] = _mod
+_spec.loader.exec_module(_mod)
 
 if __name__ == "__main__":
-    main()
+    _mod.main()
